handle extensions in WikiConfig.js
