+++ lib/config/WikiConfig.js
@@ -15,6 +15,8 @@ var JSUtils = require('../utils/jsutils.js').JSUtils;
 var Promise = require('../utils/promise.js');
 var Util = require('../utils/Util.js').Util;
 var ParamInfoRequest = require('../mw/ApiRequest.js').ParamInfoRequest;
+var ParsoidConfig = require('./ParsoidConfig').ParsoidConfig;
+var path = require('path');
 
 // Make sure our base config is never modified
 JSUtils.deepFreeze(baseConfig);
@@ -134,7 +136,25 @@ function WikiConfig(parsoidConfig, resultConf, prefix) {
 	// The interwiki prefix
 	this.iwp = prefix || "";
 
-	var mwApiConf = parsoidConfig.mwApiMap.get({ prefix, isGamepedia: general.gamepedia });
+	var mwApiConf = parsoidConfig.mwApiMap.get(prefix);
+
+	console.log('parsoid log general', JSON.stringify(general));
+	const isGamepedia = general.gamepedia;
+	const allNativeExtensions = [];
+	ParsoidConfig._collectExtensions(
+		allNativeExtensions,
+		path.resolve(__dirname, '../ext'),
+		true /* don't require a 'parsoid' subdirectory */
+	);
+	let filteredExtensions = [];
+	if (isGamepedia === true) {
+		filteredExtensions = allNativeExtensions;
+	} else {
+		filteredExtensions = allNativeExtensions.filter(function(extension) {
+			return !(extension.name === 'Gallery');
+		});
+	}
+	mwApiConf.extensions = filteredExtensions;
 
 	// The URI for api.php on this wiki.
 	this.apiURI = mwApiConf.uri;
@@ -624,7 +644,6 @@ function WikiConfig(parsoidConfig, resultConf, prefix) {
 			return FromHTML.serializeDOM(env, body, useSelser);
 		},
 	});
-	console.log('parsoid log extensions isGamepedia', JSON.stringify(general));
 	mwApiConf.extensions.forEach(config => this.registerExtension(config));
 
 	// Somewhat annoyingly, although LanguageConversion is turned on by
