+++ lib/config/WikiaApiMap.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var ParsoidConfig = require('./ParsoidConfig').ParsoidConfig;
-var path = require('path');
+// var ParsoidConfig = require('./ParsoidConfig').ParsoidConfig;
+// var path = require('path');
 
 function WikiaApiMap() {
 	this.apiMap = new Map();
@@ -14,48 +14,48 @@ WikiaApiMap.prototype.has = function(key) {
 WikiaApiMap.prototype.get = function(key) {
 	const stack = new Error().stack;
 
-	let isGamepedia = false;
+	// let isGamepedia = false;
 
 	const parsoidLog = {
-		origKey: key,
-		strippedKey: key,
-		isGamepedia,
-		isExtensionsWithoutGallery: true,
+		key,
+		// strippedKey: key,
+		// isGamepedia,
+		// isExtensionsWithoutGallery: true,
 		stack,
 		// env: '',
 		// keyAfterEnv: '',
 	};
 
-	if (key.prefix) {
-		isGamepedia = key.isGamepedia;
-		parsoidLog.isGamepedia = isGamepedia;
-		key = key.prefix;
-		parsoidLog.strippedKey = key;
-	}
-
-	// if (this.apiMap.has(key)) {
-	// 	return this.apiMap.get(key);
+	// if (key.prefix) {
+	// 	isGamepedia = key.isGamepedia;
+	// 	parsoidLog.isGamepedia = isGamepedia;
+	// 	key = key.prefix;
+	// 	parsoidLog.strippedKey = key;
 	// }
 
-	const allNativeExtensions = [];
+	if (this.apiMap.has(key)) {
+		return this.apiMap.get(key);
+	}
 
-	ParsoidConfig._collectExtensions(
-		allNativeExtensions,
-		path.resolve(__dirname, '../ext'),
-		true /* don't require a 'parsoid' subdirectory */
-	);
+	// const allNativeExtensions = [];
 
-	let extensions = [];
+	// ParsoidConfig._collectExtensions(
+	// 	allNativeExtensions,
+	// 	path.resolve(__dirname, '../ext'),
+	// 	true /* don't require a 'parsoid' subdirectory */
+	// );
+
+	// let extensions = [];
 
 	// Disable native gallery handling for wikis other then on gamepedia
-	if (isGamepedia === true) {
-		extensions = allNativeExtensions;
-		parsoidLog.isExtensionsWithoutGallery = false;
-	} else {
-		extensions = allNativeExtensions.filter(function(extension) {
-			return !(extension.name === 'Gallery');
-		});
-	}
+	// if (isGamepedia === true) {
+	// 	extensions = allNativeExtensions;
+	// 	parsoidLog.isExtensionsWithoutGallery = false;
+	// } else {
+	// 	extensions = allNativeExtensions.filter(function(extension) {
+	// 		return !(extension.name === 'Gallery');
+	// 	});
+	// }
 
 	// const envMatch = /^.*\.(verify|preview|sandbox[^\.]*)\.(fandom\.com|gamepedia\.com|wikia\.(com|org))/.exec(key);
 	// if (envMatch) {
@@ -70,7 +70,7 @@ WikiaApiMap.prototype.get = function(key) {
 	return {
 		uri: 'http://' + key + '/api.php',
 		domain: encodeURIComponent(key),
-		extensions,
+		// extensions,
 	};
 };
